refactor: 1주차 피드백을 바탕으로 리팩토링
