+++ api/build.gradle
@@ -1,0 +1,9 @@
+bootJar.enabled = true
+
+jar.enabled = false
+
+dependencies {
+    implementation project(':domain')
+
+    implementation "org.springframework.boot:spring-boot-starter-web"
+}
