+++ api/http/movie.http
@@ -1,0 +1,3 @@
+### 1. 상영중인 영화 조회
+GET localhost:8080/v1/movies/now-showing
+Content-Type: application/json
