+++ api/src/main/java/com/example/Hang99Application.java
@@ -1,0 +1,24 @@
+package com.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@EnableJpaAuditing
+@SpringBootApplication
+public class Hang99Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Hang99Application.class, args);
+    }
+
+    @Bean
+    public AuditorAware<String> auditorProvider(){
+        return () -> Optional.of("SYSTEM");
+    }
+}
