+++ api/src/main/java/com/example/movie/application/dto/ScreeningTimeDetail.java
@@ -1,0 +1,9 @@
+package com.example.movie.application.dto;
+
+import java.time.LocalDateTime;
+
+public record ScreeningTimeDetail(
+        LocalDateTime startAt,
+        LocalDateTime endAt
+) {
+}
