+++ api/src/main/java/com/example/movie/application/dto/ScreeningsDetail.java
@@ -1,0 +1,10 @@
+package com.example.movie.application.dto;
+
+import java.util.List;
+
+public record ScreeningsDetail(
+        Long theaterId,
+        String theater,
+        List<ScreeningTimeDetail> screeningTimes
+) {
+}
