+++ api/src/main/java/com/example/response/BaseErrorResponse.java
