+++ domain/build.gradle
@@ -1,0 +1,8 @@
+bootJar.enabled = false
+
+jar.enabled = true
+
+dependencies {
+    runtimeOnly 'com.mysql:mysql-connector-j'
+    api("org.springframework.boot:spring-boot-starter-data-jpa")
+}
