+++ domain/src/main/java/com/example/BaseEntity.java
@@ -1,0 +1,37 @@
+package com.example;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Data;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Data
+@EntityListeners(AuditingEntityListener.class)
+@MappedSuperclass
+public abstract class BaseEntity {
+
+    @CreatedDate
+    @Column(updatable = false)
+    private LocalDateTime createAt;
+
+
+    @LastModifiedDate
+    private LocalDateTime modifyAt;
+
+
+    @CreatedBy
+    @Column(updatable = false)
+    private String createBy;
+
+
+    @LastModifiedBy
+    private String modifyBy;
+
+}
