+++ domain/src/main/java/com/example/movie/entity/movie/Genre.java
@@ -1,0 +1,5 @@
+package com.example.movie.entity.movie;
+
+public enum Genre {
+    ACTION, ROMANCE, HORROR, SF
+}
