+++ domain/src/main/java/com/example/movie/entity/movie/Grade.java
@@ -1,0 +1,5 @@
+package com.example.movie.entity.movie;
+
+public enum Grade {
+    ALL_AGE, FROM_12_AGE, FROM_15_AGE, FROM_19_AGE, RESTRICTED
+}
