+++ domain/src/main/java/com/example/movie/entity/movie/Screening.java
@@ -1,0 +1,26 @@
+package com.example.movie.entity.movie;
+
+import com.example.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+import java.time.LocalDateTime;
+
+@Entity
+public class Screening extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long movieId;
+
+    private Long theaterId;
+
+    private LocalDateTime startAt;
+
+    private LocalDateTime endAt;
+
+}
