+++ domain/src/main/java/com/example/movie/entity/theater/Seat.java
@@ -1,0 +1,24 @@
+package com.example.movie.entity.theater;
+
+import com.example.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import org.springframework.lang.Nullable;
+
+@Entity
+public class Seat extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long theaterId;
+
+    @Nullable
+    private Long userId;
+
+    private String position;
+
+}
