+++ domain/src/main/java/com/example/movie/entity/theater/Theater.java
@@ -1,0 +1,18 @@
+package com.example.movie.entity.theater;
+
+import com.example.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+@Entity
+public class Theater extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+}
