+++ domain/src/main/java/com/example/reservation/entity/Reservation.java
@@ -1,0 +1,24 @@
+package com.example.reservation.entity;
+
+import com.example.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+import java.time.LocalDateTime;
+
+@Entity
+public class Reservation extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long seatId;
+
+    private Long userId;
+
+    private LocalDateTime reserveAt;
+
+}
