--- module-application/build.gradle
@@ -1,9 +1,0 @@
-bootJar.enabled = false
-
-jar.enabled = true
-
-dependencies {
-    implementation project(':module-core')
-    implementation project(':module-domain')
-    implementation project(':module-infra')
-}
