--- module-application/src/main/java/com/example/movie/response/ScreeningsDetail.java
@@ -1,9 +1,0 @@
-package com.example.movie.response;
-
-import java.util.List;
-
-public record ScreeningsDetail (
-        String theater,
-        List<String>screeningTime
-) {
-}
