--- module-core/build.gradle
@@ -1,12 +1,0 @@
-bootJar.enabled = false
-
-jar.enabled = true
-
-dependencies {
-    implementation "com.fasterxml.jackson.core:jackson-core:2.9.9"
-    implementation "com.fasterxml.jackson.core:jackson-annotations:2.9.9"
-    implementation "com.fasterxml.jackson.core:jackson-databind:2.9.9"
-    implementation "com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.9.9"
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-}
