--- module-domain/build.gradle
@@ -1,8 +1,0 @@
-bootJar.enabled = false
-
-jar.enabled = true
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation project(':module-core')
-}
