--- module-domain/src/main/java/com/example/entity/BaseEntity.java
@@ -1,35 +1,0 @@
-package com.example.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdDate;
-
-
-    @LastModifiedDate
-    private LocalDateTime lastModifiedDate;
-
-
-    @CreatedBy
-    @Column(updatable = false)
-    private String createdBy;
-
-
-    @LastModifiedBy
-    private String lastModifiedBy;
-
-}
