--- module-domain/src/main/java/com/example/entity/movie/Genre.java
@@ -1,5 +1,0 @@
-package com.example.entity.movie;
-
-public enum Genre {
-    액션, 로맨스, 호러, SF
-}
