--- module-domain/src/main/java/com/example/entity/movie/Grade.java
@@ -1,5 +1,0 @@
-package com.example.entity.movie;
-
-public enum Grade {
-    ALL_AGE, FROM_12_AGE, FROM_15_AGE, FROM_19_AGE, RESTRICTED
-}
