--- module-domain/src/main/java/com/example/entity/movie/Movie.java
@@ -1,28 +1,0 @@
-package com.example.entity.movie;
-
-import com.example.entity.BaseEntity;
-import jakarta.persistence.*;
-
-import java.time.LocalDate;
-
-@Entity
-public class Movie extends BaseEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String thumbnail;
-
-    @Enumerated(EnumType.STRING)
-    private Genre genre;
-
-    @Enumerated(EnumType.STRING)
-    private Grade grade;
-
-    private LocalDate releaseDate;
-
-    private Long runningTime;
-
-}
