--- module-domain/src/main/java/com/example/entity/movie/Screening.java
@@ -1,26 +1,0 @@
-package com.example.entity.movie;
-
-import com.example.entity.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-import java.time.LocalDateTime;
-
-@Entity
-public class Screening extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Long movieId;
-
-    private Long theaterId;
-
-    private LocalDateTime startTime;
-
-    private LocalDateTime endTime;
-
-}
