--- module-domain/src/main/java/com/example/entity/movie/Theater.java
@@ -1,18 +1,0 @@
-package com.example.entity.movie;
-
-import com.example.entity.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Theater extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-}
