--- module-infra/build.gradle
@@ -1,10 +1,0 @@
-bootJar.enabled = false
-
-jar.enabled = true
-
-dependencies {
-    api("org.springframework.boot:spring-boot-starter-data-jpa")
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    implementation project(':module-core')
-    implementation project(':module-domain')
-}
