--- module-infra/src/main/resources/application.yml
@@ -1,17 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3305/hanghae99
-    username: root
-    password: 1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  sql:
-    init:
-      platform: mysql
-      mode: always
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-        show_sql: true
