--- module-presentation/build.gradle
@@ -1,12 +1,0 @@
-bootJar.enabled = false
-
-jar.enabled = true
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation project(':module-core')
-    implementation project(':module-infra')
-    implementation project(':module-domain')
-    implementation project(':module-application')
-}
