+++ settings.gradle
@@ -1,7 +1,4 @@
 rootProject.name = 'multi-modules'
-include 'module-presentation'
-include 'module-application'
-include 'module-domain'
-include 'module-infra'
-include 'module-core'
+include 'api'
+include 'domain'
